+++ src/main/java/com/eazy/pay/dao/FaqRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface FaqRepository extends JpaRepository<Faq, Integer> {
+public interface FaqRepository extends JpaRepository<Faq, Long> {
     List<Faq> findAll();
 }
